--- Entry.cpp
@@ -1,7 +1,0 @@
-#include "Performance.h"
-
-int main(int argc, char **argv)
-{
-    Performance project(argc, argv);
-    return project.run();
-}
